[TASK-211] <<fixed>> rename MainBanner to Goal5LearningBanner
