+++ itp-front/src/views/AnswerQuestionPage.vue
@@ -1,7 +1,7 @@
 <template>
     <div id="AnswerQuestion">
       <Header></Header>
-      <MainBanner></MainBanner>
+      <Goal5LearningBanner></Goal5LearningBanner>
       <div class="sidebar-container">
         <SideBar class="sidebar"></SideBar>
         <div class="questions-detail">
@@ -19,7 +19,7 @@
   </template>
   
   <script>
-  import MainBanner from '../components/MainBannerLearn.vue';
+  import Goal5LearningBanner from '../components/Goal5LearningBanner.vue';
   import MainFooter from '../components/MainFooter.vue';
   import SideBar from '../components/SideBar.vue';
   import QuestionDetail from '../components/QuestionDetail.vue';
@@ -29,7 +29,7 @@
     name: 'AnswerQuestion',
     components: {
       Header,
-      MainBanner,
+      Goal5LearningBanner,
       MainFooter,
       SideBar,
       QuestionDetail
